Added more explanation and images to Data Preparation section
Fixed spelling errors. Added an explanation for the pipeline failing because the data source is missing (with screenshot). Added some stats calculation to the Data manipulation section which helps justify the code which removes outliers. Added some text to explain why we generate scatter plots in the data visualization section. Separated the data splitting and saving into its own section to avoid confusion with the data visualization section. Added an explanation of model monitor and why it requires a baseline data file to be saved at this stage.