Updated plots and added explanations to the Inspect Training Job section
Added an internal link to the end of the Inspect Training Job section, which allows users to quickly navigate back to this section if they want to compare the test results with the validation results. Added explanations and a screenshot to the Test Dev Deployment section. Changed the scatter plot from a pandas plot to a seaborn plot for consistency with previous sections. Changed the scatter plot to show all results instead of just the tail of a sorted list of results, which was causing confusion.