Expanded Approve Deployment to Production section, added explanation to Test Production Deployment section
Renamed the Approve Prod Deployment section to Approve Deployment to Production for clarity. Expanded on the text explanations in the Approve Deployment to Production section. Added more explanations and a screenshot to the Test Production Deployment section. One TODO needs to be replaced in a future commit when I better understand what happens in that step.