Added code and permissions for deleting monitoring schedules
Added permissions to the SageMaker role for listing monitoring schedules and deleting monitoring schedules. This is required for cleaning up because CloudFormation fails to delete the monitoring schedule automatically. Also added code to the Clean Up section of the notebook for deleting the monitoring schedules attached to the production endpoint. Removed stack names from the Clean Up section, which I believe was legacy.