Updated architecture diagram, added explanations to overview
Significantly expanded the Overview section with a more detailed architecture diagram and explanations for each step of the process. Added an Info notice about the prerequisites for completing this demo. Added an explanation of what we mean by a ‘safe’ deployment and included links to documentation on canary deployment, least privilege.