Various small adjustments (see description)
Updated the alt text to be relevant for each image.
Added clarification on the types of notifications sent by SNS.
Added an explanation for the NaN results for the baseline job, and added a link to the section where the baseline output is explored.
Removed my temporary fix for the workflow ARN problem.
Added an explanation for the CodeDeploy component of the production deployment.
Rewrote the explanations in the Test REST API Application section.
Timed every step of the notebook and added the average waiting times to the explanations.