Complete rewrite of the Jupyter notebook for better workshop support (#24)
* Removed histogram and replaced matplotlib with seaborn for scatter plots

* Added more explanation and images to Data Preparation section

Fixed spelling errors. Added an explanation for the pipeline failing because the data source is missing (with screenshot). Added some stats calculation to the Data manipulation section which helps justify the code which removes outliers. Added some text to explain why we generate scatter plots in the data visualization section. Separated the data splitting and saving into its own section to avoid confusion with the data visualization section. Added an explanation of model monitor and why it requires a baseline data file to be saved at this stage.

* Added more explanation to the Start model build section

Added text to explain that the environment variables are set by a lifecycle configuration script. Added an explanation of the ZIP file required as the S3 source for CodePipeline. Clarified that the workflow notebook is optional and added a note about its purpose.

* Small fixes in Start Model Build and explanations added to Inspect Training Job

Fixed capitalization of Start Model Build to be consistent with the other section headers of the same level. Cleared output which I accidentally added in my previous commit. Removed a temporary workflow pipeline arn fix which I accidentally submitted in my previous commit. Renamed the Wait for Training Job section to Inspect Training Job to better reflect what happens in the section. Added explanations and screenshots to the Inspect Training Job section.

* Updated plots and added explanations to the Inspect Training Job section

Added an internal link to the end of the Inspect Training Job section, which allows users to quickly navigate back to this section if they want to compare the test results with the validation results. Added explanations and a screenshot to the Test Dev Deployment section. Changed the scatter plot from a pandas plot to a seaborn plot for consistency with previous sections. Changed the scatter plot to show all results instead of just the tail of a sorted list of results, which was causing confusion.

* Expanded Approve Deployment to Production section, added explanation to Test Production Deployment section

Renamed the Approve Prod Deployment section to Approve Deployment to Production for clarity. Expanded on the text explanations in the Approve Deployment to Production section. Added more explanations and a screenshot to the Test Production Deployment section. One TODO needs to be replaced in a future commit when I better understand what happens in that step.

* Added code and permissions for deleting monitoring schedules

Added permissions to the SageMaker role for listing monitoring schedules and deleting monitoring schedules. This is required for cleaning up because CloudFormation fails to delete the monitoring schedule automatically. Also added code to the Clean Up section of the notebook for deleting the monitoring schedules attached to the production endpoint. Removed stack names from the Clean Up section, which I believe was legacy.

* Moved sagemaker-custom-resource to end of the Clean Up section because it gets recreated otherwise

* Updated architecture diagram, added explanations to overview

Significantly expanded the Overview section with a more detailed architecture diagram and explanations for each step of the process. Added an Info notice about the prerequisites for completing this demo. Added an explanation of what we mean by a ‘safe’ deployment and included links to documentation on canary deployment, least privilege.

* Added more explanation and a screenshot to Model Monitor section

Expanded the Inspect Model Monitor section with more explanations and links, as well as a screenshot of the CloudWatch alarm which users can expect to see.

* Added more explanation and screenshots to CloudWatch Monitoring section

Expanded the CloudWatch Monitoring section with more explanations, as well as screenshots of the CloudWatch canaries and dashboard pages.

* Removed monitoring schedule code from Clean Up section

Turns out that CloudFormation has trouble deleting the monitoring schedule attached to the production endpoint due to a permissions error. This means we won't need the code for removing the monitoring schedule in the notebook itself. We only have to fix the sagemaker-custom-resource IAM role.

* Fixed broken markdown cell in mlops notebook

* Various small adjustments (see description)

Updated the alt text to be relevant for each image.
Added clarification on the types of notifications sent by SNS.
Added an explanation for the NaN results for the baseline job, and added a link to the section where the baseline output is explored.
Removed my temporary fix for the workflow ARN problem.
Added an explanation for the CodeDeploy component of the production deployment.
Rewrote the explanations in the Test REST API Application section.
Timed every step of the notebook and added the average waiting times to the explanations.

* Small changes based on feedback

* Added more cells to view code files while the participant waits (i.e. during model monitoring schedule and during endpoint deployment)
* Updated the alarm screenshot and added another screenshot for traffic shifting
* Clarified the explanations in several places based on feedback from Julian